+++ src/data/audioStream.h
@@ -1,3 +1,4 @@
+#include "util.h"
 #include "types.h"
 
 #pragma once
@@ -6,20 +7,20 @@ struct Blob;
 
 typedef struct AudioStream {
   Ref ref;
-  int bitDepth;
-  int channelCount;
-  int sampleRate;
-  int samples;
-  int bufferSize;
+  u32 bitDepth;
+  u32 channelCount;
+  u32 sampleRate;
+  usize samples;
+  u32 bufferSize;
   void* buffer;
   void* decoder;
   struct Blob* blob;
 } AudioStream;
 
-AudioStream* lovrAudioStreamInit(AudioStream* stream, struct Blob* blob, int bufferSize);
+AudioStream* lovrAudioStreamInit(AudioStream* stream, struct Blob* blob, usize bufferSize);
 #define lovrAudioStreamCreate(...) lovrAudioStreamInit(lovrAlloc(AudioStream), __VA_ARGS__)
 void lovrAudioStreamDestroy(void* ref);
-int lovrAudioStreamDecode(AudioStream* stream, short* destination, int size);
+usize lovrAudioStreamDecode(AudioStream* stream, i16* destination, usize size);
 void lovrAudioStreamRewind(AudioStream* stream);
-void lovrAudioStreamSeek(AudioStream* stream, int sample);
-int lovrAudioStreamTell(AudioStream* stream);
+void lovrAudioStreamSeek(AudioStream* stream, usize sample);
+usize lovrAudioStreamTell(AudioStream* stream);
